@@ -1,29 +1,45 @@
 {
   "private": true,
   "scripts": {
-    "dev": "npm run development",
-    "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "watch": "npm run development -- --watch",
-    "watch-poll": "npm run watch -- --watch-poll",
-    "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "prod": "npm run production",
-    "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+    "dev": "mix",
+    "development": "mix",
+    "watch": "mix watch",
+    "watch-poll": "mix watch -- --watch-options-poll=1000",
+    "hot": "mix watch --hot",
+    "prod": "mix --production",
+    "production": "mix --production",
+    "lint": "eslint --ext .js,.vue resources && stylelint \"resources/themes/*.scss\"",
+    "jslint": "eslint --ext .js,.vue resources"
   },
   "devDependencies": {
-    "axios": "^0.19",
-    "bootstrap": "^4.3.1",
-    "cross-env": "^5.1",
-    "jquery": "^3.4.1",
-    "laravel-mix": "^4.0.7",
-    "lodash": "^4.17.13",
-    "popper.js": "^1.12",
-    "resolve-url-loader": "^2.3.1",
-    "sass": "^1.15.2",
-    "sass-loader": "^7.1.0",
-    "vue": "^2.5.17",
-    "vue-router": "^3.1.3",
-    "vue-template-compiler": "^2.6.10",
-    "vuex": "^3.1.1",
-    "@deveodk/vue-toastr": "^1.1.0"
+    "@babel/eslint-parser": "^7.17.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.35",
+    "@fortawesome/free-brands-svg-icons": "^6",
+    "@fortawesome/free-regular-svg-icons": "^6",
+    "@fortawesome/free-solid-svg-icons": "^6",
+    "@fortawesome/vue-fontawesome": "^2.0.6",
+    "@popperjs/core": "^2.11.5",
+    "axios": "^1.7.2",
+    "bootstrap": "~5.3.0",
+    "cash-dom": "8.1.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.26.0",
+    "eslint-plugin-vue": "^7.20.0",
+    "frappe-charts": "^1.5.8",
+    "laravel-mix": "^6.0.43",
+    "linkify-string": "^4.0.0",
+    "mediasoup-client": "^3.7.8",
+    "postcss": "^8.4.12",
+    "resolve-url-loader": "^5.0.0",
+    "sass": "^1.50.1",
+    "sass-loader": "^8.0.0",
+    "socket.io-client": "^4.6.1",
+    "stylelint": "^16.6.1",
+    "stylelint-config-standard": "^36.0.0",
+    "vue": "^2.6.12",
+    "vue-i18n": "^8.27.1",
+    "vue-loader": "^15.10.0",
+    "vue-router": "^3.5.3",
+    "vue-template-compiler": "^2.6.12"
   }
 }
