@@ -4,20 +4,27 @@ namespace App\Jobs;
 
 use App\Mail\SignupVerification;
 use App\SignupCode;
-
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
-use Illuminate\Support\Facades\Mail;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Queue\InteractsWithQueue;
 
 class SignupVerificationEmail implements ShouldQueue
 {
-    use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
+    use Dispatchable;
+    use InteractsWithQueue;
+    use Queueable;
+    use SerializesModels;
 
     /** @var int The number of times the job may be attempted. */
-    public $tries = 2;
+    public $tries = 3;
+
+    /** @var int The number of seconds to wait before retrying the job. */
+    public $backoff = 10;
+
+    /** @var bool Delete the job if its models no longer exist. */
+    public $deleteWhenMissingModels = true;
 
     /** @var SignupCode Signup verification code object */
     protected $code;
@@ -36,23 +43,16 @@ class SignupVerificationEmail implements ShouldQueue
     }
 
     /**
-     * Determine the time at which the job should timeout.
-     *
-     * @return \DateTime
-     */
-    public function retryUntil()
-    {
-        // FIXME: I think it does not make sense to continue trying after 1 hour
-        return now()->addHours(1);
-    }
-
-    /**
      * Execute the job.
      *
      * @return void
      */
     public function handle()
     {
-        Mail::to($this->code->data['email'])->send(new SignupVerification($this->code));
+        \App\Mail\Helper::sendMail(
+            new SignupVerification($this->code),
+            $this->code->tenant_id,
+            ['to' => $this->code->email]
+        );
     }
 }
