@@ -3,6 +3,7 @@
 namespace App\Observers;
 
 use App\Entitlement;
+use Carbon\Carbon;
 
 /**
  * This is an observer for the Entitlement model definition.
@@ -19,30 +20,17 @@ class EntitlementObserver
      *
      * @param Entitlement $entitlement The entitlement being created.
      *
-     * @return void
+     * @return bool
      */
-    public function creating(Entitlement $entitlement)
+    public function creating(Entitlement $entitlement): bool
     {
-        $entitlement->{$entitlement->getKeyName()} = \App\Utils::uuidStr();
-
         // can't dispatch job here because it'll fail serialization
 
         // Make sure the owner is at least a controller on the wallet
-        $owner = \App\User::find($entitlement->owner_id);
         $wallet = \App\Wallet::find($entitlement->wallet_id);
 
-        if (!$owner) {
+        if (!$wallet || !$wallet->owner) {
             return false;
-        }
-
-        if (!$wallet) {
-            return false;
-        }
-
-        if (!$wallet->owner() == $owner) {
-            if (!$wallet->controllers->contains($owner)) {
-                return false;
-            }
         }
 
         $sku = \App\Sku::find($entitlement->sku_id);
@@ -51,6 +39,54 @@ class EntitlementObserver
             return false;
         }
 
-        $wallet->debit($sku->cost);
+        $result = $sku->handler_class::preReq($entitlement, $wallet->owner);
+
+        if (!$result) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * Handle the entitlement "created" event.
+     *
+     * @param \App\Entitlement $entitlement The entitlement.
+     *
+     * @return void
+     */
+    public function created(Entitlement $entitlement)
+    {
+        $entitlement->createTransaction(\App\Transaction::ENTITLEMENT_CREATED);
+
+        $entitlement->sku->handler_class::entitlementCreated($entitlement);
+    }
+
+    /**
+     * Handle the entitlement "deleted" event.
+     *
+     * @param \App\Entitlement $entitlement The entitlement.
+     *
+     * @return void
+     */
+    public function deleted(Entitlement $entitlement)
+    {
+        $entitlement->createTransaction(\App\Transaction::ENTITLEMENT_DELETED);
+
+        $entitlement->sku->handler_class::entitlementDeleted($entitlement);
+    }
+
+    /**
+     * Handle the entitlement "deleting" event.
+     *
+     * @param \App\Entitlement $entitlement The entitlement.
+     *
+     * @return void
+     */
+    public function deleting(Entitlement $entitlement)
+    {
+        // Disable updating of updated_at column on delete, we need it unchanged to later
+        // charge the wallet for the uncharged period before the entitlement has been deleted
+        $entitlement->timestamps = false;
     }
 }
