@@ -4,6 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
+// phpcs:ignore
 class WalletControllers extends Migration
 {
     /**
@@ -19,22 +20,12 @@ class WalletControllers extends Migration
                 $table->bigIncrements('id');
                 $table->bigInteger('user_id');
                 $table->string('wallet_id', 36);
-                $table->timestamps();
-            }
-        );
+                $table->timestamp('created_at')->useCurrent();
+                $table->timestamp('updated_at')->useCurrent();
 
-        Schema::table(
-            'user_accounts',
-            function (Blueprint $table) {
                 $table->unique(['user_id', 'wallet_id']);
-
-                $table->foreign('user_id')
-                    ->references('id')->on('users')
-                    ->onDelete('cascade');
-
-                $table->foreign('wallet_id')
-                    ->references('id')->on('wallets')
-                    ->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('wallet_id')->references('id')->on('wallets')->onDelete('cascade');
             }
         );
     }
