@@ -13,7 +13,15 @@ return [
     |
     */
 
-    'failed' => 'These credentials do not match our records.',
+    'failed' => 'Invalid username or password.',
+    'password' => 'The provided password is incorrect.',
     'throttle' => 'Too many login attempts. Please try again in :seconds seconds.',
+    'logoutsuccess' => 'Successfully logged out.',
+
+    'error.password' => "Invalid password",
+    'error.geolocation' => "Country code mismatch",
+    'error.nofound' => "User not found",
+    'error.2fa' => "Second factor failure",
+    'error.2fa-generic' => "Second factor failure",
 
 ];
