@@ -1,4 +1,3 @@
-const mix = require('laravel-mix');
 
 /*
  |--------------------------------------------------------------------------
@@ -11,5 +10,33 @@ const mix = require('laravel-mix');
  |
  */
 
-mix.js('resources/js/app.js', 'public/js')
-    .sass('resources/sass/app.scss', 'public/css');
+const { spawn } = require('child_process');
+const glob = require('glob');
+const mix = require('laravel-mix');
+
+mix.options({
+    vue: {
+        compilerOptions: {
+            whitespace: 'condense'
+        }
+    }
+})
+
+// Prepare some resources before compilation
+mix.before(() => {
+    spawn('php', ['resources/build/before.php'], { stdio: 'inherit' })
+})
+
+// Compile the Vue/js resources
+mix.js('resources/js/user/app.js', 'public/js/user.js')
+    .js('resources/js/admin/app.js', 'public/js/admin.js')
+    .js('resources/js/reseller/app.js', 'public/js/reseller.js')
+    .vue()
+
+// Compile the themes/css resources
+glob.sync('resources/themes/*/', {}).forEach(fromDir => {
+    const toDir = fromDir.replace('resources/themes/', 'public/themes/')
+
+    mix.sass(fromDir + 'app.scss', toDir)
+        .sass(fromDir + 'document.scss', toDir);
+})
